+++ packages/lib-sourcify/src/lib/SourcifyChain.ts
@@ -6,7 +6,7 @@ import {
   getAddress,
 } from 'ethers';
 import { Chain, SourcifyChainExtension } from './types';
-import { logInfo, logWarn } from './logger';
+import { logError, logInfo, logWarn } from './logger';
 
 const RPC_TIMEOUT = process.env.RPC_TIMEOUT
   ? parseInt(process.env.RPC_TIMEOUT)
@@ -27,7 +27,6 @@ export default class SourcifyChain {
   chainId: number;
   rpc: Array<string | FetchRequest>;
   supported: boolean;
-  monitored: boolean;
   contractFetchAddress?: string | undefined;
   graphQLFetchAddress?: string | undefined;
   txRegex?: string[] | undefined;
@@ -39,7 +38,6 @@ export default class SourcifyChain {
     this.chainId = sourcifyChainObj.chainId;
     this.rpc = sourcifyChainObj.rpc;
     this.supported = sourcifyChainObj.supported;
-    this.monitored = sourcifyChainObj.monitored;
     this.contractFetchAddress = sourcifyChainObj.contractFetchAddress;
     this.graphQLFetchAddress = sourcifyChainObj.graphQLFetchAddress;
     this.txRegex = sourcifyChainObj.txRegex;
@@ -84,7 +82,10 @@ export default class SourcifyChain {
 
   rejectInMs = (ms: number, host?: string) =>
     new Promise<never>((_resolve, reject) => {
-      setTimeout(() => reject(`RPC ${host} took too long to respond`), ms);
+      setTimeout(
+        () => reject(new Error(`RPC ${host} took too long to respond`)),
+        ms
+      );
     });
 
   getTx = async (creatorTxHash: string) => {
@@ -155,7 +156,7 @@ export default class SourcifyChain {
       } catch (err) {
         if (err instanceof Error) {
           logWarn(
-            `Can't fetch bytecode from RPC ${provider.url} and chain ${this.chainId}`
+            `Can't fetch ${address}'s bytecode from RPC ${provider.url} and chain ${this.chainId}`
           );
           continue;
         } else {
@@ -180,26 +181,26 @@ export default class SourcifyChain {
           provider.getBlock(blockNumber, preFetchTxs),
           this.rejectInMs(RPC_TIMEOUT, provider.url),
         ]);
-        logInfo(
-          'Block fetched from ' +
-            provider.url +
-            ' for ' +
-            blockNumber +
-            ' on chain ' +
-            this.chainId
-        );
-        return block;
-      } catch (err) {
-        if (err instanceof Error) {
-          logWarn(
-            `Can't fetch block ${blockNumber} from RPC ${provider.url} and chain ${this.chainId}, error: ${err.message}`
+        if (block) {
+          logInfo(
+            `Block ${blockNumber} fetched from ${provider.url} on chain ${this.chainId}`
           );
-          continue;
         } else {
-          throw err;
+          logInfo(
+            `Block ${blockNumber} not published yet on ${this.chainId} from ${provider.url}`
+          );
         }
+        return block;
+      } catch (err: any) {
+        logWarn(
+          `Can't fetch block ${blockNumber} from RPC ${provider.url} and chain ${this.chainId}, error: ${err.message}`
+        );
+        continue;
       }
     }
+    logError(
+      `None of the RPCs responded fetching block ${blockNumber} on chain ${this.chainId}`
+    );
     throw new Error(
       'None of the RPCs responded fetching block ' +
         blockNumber +