+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [sourcify-server@1.1.1] - 2023-10-03
+
+- Remove monitor code from the server directory `src/`
+- Update chains.json
+- Renaming `ALCHEMY_ID` and `INFURA_ID` to `..._API_KEY`
+- Remove the `monitored` field from `sourcify-chains.ts`
+- New chains:
+  - Ethereum Holesky Testnet (17000)
+  - PulseChain Mainnet (369)
+  - Mind Smart Chain Mainnet (9996)
+  - Mind Smart Chain Testnet (9977)
+  - Shrapnel Testnet (2038)
+  - Shrapnel Subnet (2044)
+  - Arthera Testnet (10243)
+  - Core Blockchain Mainnet (1116)
+  - Q Mainnet (35441)
+  - Q Testnet (35443)
+
 ## [sourcify-server@1.1.0] - 2023-09-04
 
 - Updated lerna to `7.1.5`